requirements.txt
transformers
torch
feedparser
beautifulsoup4
requests
newsapi-python
python-dotenv
schedule
jinja2
yfinance
flask
